requirements.txt
polib
deepl
tqdm
requests
python-dotenv
